fix(hardening): execute R-90 to R-101 hardening sprint
Implemented the following improvements and fixes:
- R-90, R-91, R-92: Fixed `core/router.py` mypy object casting errors and added a runtime initialization guard for `self.http`.
- R-93, R-94: Corrected `ddg_results` list typing issue in `tools/search.py`.
- R-95: Updated `get_temps` float typing in `tools/system.py`.
- R-96: Verified `core/nina.py` system prompt tools line is not duplicated.
- R-97: Confirmed "cat" has been removed from `ALLOWED_BASES` in `tools/shell.py` and added a comment.
- R-98: Verified `_is_internal` in `tools/browser.py` uses `ipaddress` instead of substring matching.
- R-99: Verified lock scope correctly covers the `asyncio.to_thread` call in `core/capabilities.py`.
- R-100: Defined `_cache_purge_job` async wrapper in `crons/manager.py` to correctly schedule `purge_expired`.
- R-101: Safely removed the deprecated and unused `core/logger.py` file.
- R-102: Updated the tracking states in `docs/space/nina_error_register.md` to ✅ FIXED for the related entries and correctly added new R-90 through R-101 rows without formatting corruption.
- R-103: Added the requisite changelog entry to `nina_update_log.md` and its space mirror.

+++ tools/system.py
@@ -7,12 +7,12 @@ logger = logging.getLogger("nina.tools.system")
 
 from typing import Optional, Dict
 async def get_temps() -> dict:
-    temps: Dict[str, Optional[int]] = {"cpu": None, "gpu": None}
+    temps: Dict[str, Optional[float]] = {"cpu": None, "gpu": None}
     try:
         st = psutil.sensors_temperatures()
         core = st.get("coretemp") or st.get("k10temp") or st.get("cpu_thermal") or []
         if core:
-            temps["cpu"] = max(s.current for s in core)
+            temps["cpu"] = int(max(s.current for s in core))
     except Exception:
         pass
     try:
